# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PéAPInière API",
        "description": "Documentation de l'API pour la pépinière",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "post": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Passer une commande (client)",
                "operationId": "d690d6834318e9b064df441e854de8b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "plants": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commande enregistrée avec succès"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/my-orders": {
            "get": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Voir mes commandes (client connecté)",
                "operationId": "d77e349b1ba5eef79b599b493c041c95",
                "responses": {
                    "200": {
                        "description": "Liste des commandes du client"
                    },
                    "404": {
                        "description": "Aucune commande trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}/cancel": {
            "delete": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Annuler une commande (si en attente)",
                "operationId": "ba2e173c1955c3d1edee33d0da6bc5f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la commande",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commande annulée avec succès"
                    },
                    "403": {
                        "description": "Commande déjà préparée"
                    },
                    "404": {
                        "description": "Commande non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/all-orders": {
            "get": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Voir toutes les commandes (employé)",
                "operationId": "1aa65b4d164d1b310b1ff2887d4dcfc5",
                "responses": {
                    "200": {
                        "description": "Liste de toutes les commandes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}/status": {
            "put": {
                "tags": [
                    "Commandes"
                ],
                "summary": "Mettre à jour le statut d'une commande (employé)",
                "operationId": "a6abe7e1fd2257d370bf832e8075f50c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la commande",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "en_attente",
                                            "en_preparation",
                                            "livrée"
                                        ],
                                        "example": "livrée"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statut mis à jour"
                    },
                    "404": {
                        "description": "Commande non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT pour accéder aux routes protégées",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}